+++ app/assets/stylesheets/base.scss.erb
@@ -14,14 +14,14 @@
 }
 
 .nameCounter.forTopic {
-  
+
 }
 
 #center-container {
   position:relative;
 	height:100%;
 	width:100%;
-  
+
   /* background-color:#031924; */
   color:#444;
 }
@@ -169,7 +169,7 @@
 
 .CardOnGraph .best_in_place_desc {
   display:block;
-  margin-top:2px;	
+  margin-top:2px;
   padding-right: 18px;
   margin-right: 8px;
 }
@@ -213,50 +213,18 @@
   padding: 0;
   height: 48px;
 }
-.linkItem.contributor {
-  margin-left:40px;
-  z-index:1;
-  padding:17px 16px 17px 30px;
-  position: relative;
-}
-.contributor .contributorIcon {
-  position: absolute;
-  top: 8px;
-  left: 0;
-  border-radius: 16px;
-}
 
-.contributor:hover .contributorName {
-  display: block;
-}
-
-.contributorName {
-  display: none;
-  position: absolute;
-  background: black;
-  text-align: center;
-  color: white;
-  border-radius: 2px;
-  font-family: din-regular;
-  line-height: 15px;
-  font-size: 12px;
-  padding: 3px 5px 2px;
-  white-space: nowrap;
-  margin-top: 36px;
-  margin-left: -32px;
-}
-
-.contributor div:before {
-  content: '';
-  position: absolute;
-  top: 128%;
-  left: 13px;
-  margin-top: -30px;
-  width: 0;
-  height: 0;
-  border-bottom: 4px solid #000000;
-  border-left: 5px solid transparent;
-  border-right: 5px solid transparent;
+.contributor {
+  .contributorIcon {
+    display: inline-block;
+    vertical-align: middle;
+    border-radius: 14px;
+    margin: 5px;
+  }
+  span {
+    font-family: 'din-regular', sans-serif;
+    font-size: 14px;
+  }
 }
 
 .linkItem.mapCount {
@@ -493,19 +461,21 @@ cursor: pointer;
   background-position: 0 -32px;
 }
 .permission.canEdit .minimize .expandMetacodeSelect {
-  
+
 }
 
 .CardOnGraph .metacodeImage {
-  cursor:move;
-  width:46px;
-  height:46px;
-  position:absolute;
-  left:-23px;
-  top:0;
-  background-size:46px 46px;
+  width:32px;
+  height:32px;
+  background-size:32px 32px;
   background-position:0 0;
   background-repeat:no-repeat;
+  display: inline-block;
+  vertical-align: middle;
+  margin: 5px;
+}
+.CardOnGraph .metacodeName {
+  display: inline-block;
 }
 
 #metacodeOptions {
@@ -521,7 +491,7 @@ cursor: pointer;
   white-space: nowrap;
 }
 .CardOnGraph .metacodeSelect ul {
-  position: relative; 
+  position: relative;
   line-height: 14px;
   font-size: 14px;
   font-family: helvetica, sans-serif;
@@ -623,18 +593,15 @@ background-color: #E0E0E0;
  z-index:100;
 }
 
-#embedlyLink {
-  display: none;
-}
 #embedlyLinkLoader {
   margin: 0 auto;
   width: 28px;
 }
 
 .CardOnGraph .attachments {
-  border-top: 1px solid #BDBDBD;
   width:100%;
   height:47px;
+  position: relative;
 }
 
 .attachments a {
@@ -642,7 +609,7 @@ background-color: #E0E0E0;
   overflow: hidden;
   text-overflow: ellipsis;
   display: block;
-  margin-left: 40px; 
+  margin-left: 40px;
   padding-top:9px;
   font-size: 16px;
   line-height: 16px;
@@ -652,7 +619,7 @@ background-color: #E0E0E0;
   display: inline-block;
   width: 102px;
   height: 12px;
-  text-align: left; 
+  text-align: left;
   padding: 18px 0 18px 48px;
   font-size: 12px;
   color: #9e9e9e;
@@ -752,7 +719,6 @@ font-family: 'din-regular', helvetica, sans-serif;
   -moz-border-radius-bottomright: 8px;
   -webkit-border-bottom-right-radius: 8px;
   border-bottom-right-radius: 8px;
-  display: none;
   margin: 8px;
 }
 
@@ -950,11 +916,11 @@ font-family: 'din-regular', helvetica, sans-serif;
 }
 #edit_synapse_right {
   background-image: url(<%= asset_data_uri('synapsedirectionright_sprite.png') %>);
-  right: 16px; 
+  right: 16px;
 }
 #edit_synapse_left {
   background-image: url(<%= asset_data_uri('synapsedirectionleft_sprite.png') %>);
-  right: 56px; 
+  right: 56px;
 }
 #edit_synapse_left.checked, #edit_synapse_right.checked {
   background-position: 0 -48px;
