+++ frontend/src/Metamaps/JIT.js
@@ -18,7 +18,7 @@ import Settings from './Settings'
 import Synapse from './Synapse'
 import SynapseCard from './SynapseCard'
 import Topic from './Topic'
-import TopicCard from './TopicCard'
+import TopicCard from './Views/TopicCard'
 import Util from './Util'
 import Visualize from './Visualize'
 import clipboard from 'clipboard-js'
