+++ frontend/src/Metamaps/Views/ChatView.js
@@ -49,6 +49,7 @@ const ChatView = {
     $('#' + ChatView.domId).hide()
   },
   render: () => {
+    return
     if (!Active.Map) return
     const self = ChatView
     self.mapChat = ReactDOM.render(React.createElement(MapChat, {
